tidied up the code a little

+++ scala-kata/src/main/scala/org/acmebank/katas/bankaccount/Money.scala
@@ -3,5 +3,5 @@ package org.acmebank.katas.bankaccount
 object Money {
 }
 
-case class Money(amount: Double) {
+case class Money(amount: BigDecimal) {
 }
